+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2026-07-04
+
+### Added
+
+- Documentation for the `magenpy` CLI scripts.
+- Complete testing of the CLI scripts in `test_cli.sh`.
+
+### Changed
+
+- Removed `output_dir` from `GWADataLoader` initialization.
+
+### Fixed
+
+- Fixed CLI script documentation formatting issues.
+- Fixed bugs in handling of backends in `mgp_compute_ld` and `mgp_simulate`.
+
 ## [0.2.0] - 2026-07-01
 
 ### Changed